JokeBot/pyproject.toml
[tool.poetry]
name = "smeliana"
version = "0.1.0"
description = ""
authors = ["loysee.inc"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.2"
aiogram = "^2.25.1"
transformers = "^4.27.4"
pandas = "^2.0.0"
openai = "^0.27.4"
sqlalchemy = "^2.0.9"
asyncpg = "^0.27.0"
setuptools = "^59.5.0"
direnv = "^2020.12.3"
alembic = "^1.10.3"
google-cloud-translate = "^2.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
